refactor: added school/employer name to the markers on the map
created new functions and modified existing functions.
locationFinderObj creates new object with name, and location of each school and employer.

pinPosterObj is a modified existing function, was modified to look through the new object being passed.

callback was modified to add the lng and lat to each item in object created in locationFinderObj, lng and lat from the google map textSearch

getName was created to return name of school or employer after passing lat and lng